+++ lib/castle_devise.rb
@@ -72,7 +72,8 @@ module Devise
   @@castle_hooks = {
     before_registration: true,
     after_login: true,
-    after_password_reset_request: true
+    after_password_reset_request: true,
+    profile_update: true
   }
 end