+++ lib/castle_devise/models/castle_protectable.rb
@@ -10,9 +10,12 @@ module Devise
     #   castle_hooks: configures which events trigger Castle API calls
     #     {
     #       after_login: true, # trigger risk($login) and log($login, $failed),
-    #       before_registration: true # trigger filter($registration),
-    #       after_password_reset_request: true # trigger log($password_reset_request, $succeeded)
-    #                                          # and log($password_reset_request, $failed)
+    #       before_registration: true, # trigger filter($registration)
+    #       after_password_reset_request: true, # trigger log($password_reset_request, $succeeded)
+    #                                           # and log($password_reset_request, $failed)
+    #       profile_update: true # trigger risk($profile_update, $attempted),
+    #                            # log($profile_update, $succeeded)
+    #                            # and log($profile_update, $failed)
     #     }
     module CastleProtectable
       extend ActiveSupport::Concern