+++ spec/dummy_app/app/models/user.rb
@@ -6,6 +6,7 @@ class User < ApplicationRecord
     :castle_protectable, castle_hooks: {
       after_login: true,
       after_password_reset_request: true,
-      before_registration: true
+      before_registration: true,
+      profile_update: true
     }
 end